added fixed ip addresses to example and repackaged kafka client as war

+++ kafka-experiments/integration-example1/kafka-java-client/src/test/java/org/opennms/examples/integration/exx1/manual/AlarmMessageClientTest.java
@@ -39,9 +39,11 @@ class AlarmMessageClientTest {
 		} catch (InterruptedException e) {
 			LOG.info("finisned test.");
 			return;
+		} finally {
+			if(aeclient!=null) aeclient.destroy();
 		}
 
-		aeclient.destroy();
+		
 	}
 
 }
